perf: 脚本任务执行结果拉取优化 #2919

+++ src/backend/commons/gse-sdk/src/main/java/com/tencent/bk/job/common/gse/v2/model/GetExecuteScriptResultRequest.java
@@ -65,7 +65,7 @@ public class GetExecuteScriptResultRequest extends GseReq {
         /**
          * 执行日志读取大小上限，单位byte
          */
-        private Integer limit;
+        private Long limit;
     }
 
     /**
@@ -79,7 +79,7 @@ public class GetExecuteScriptResultRequest extends GseReq {
     public void addAgentTaskQuery(ExecuteObjectGseKey executeObjectGseKey,
                                   Integer atomicTaskId,
                                   int offset,
-                                  Integer limit) {
+                                  Long limit) {
         AgentTask agentTask = new AgentTask();
         agentTask.setAgentId(executeObjectGseKey.getAgentId());
         agentTask.setContainerId(executeObjectGseKey.getContainerId());
